\name{ntfreq-methods}
\docType{methods}
\alias{ntfreq-methods}
\alias{ntfreq,DataFrame,numeric-method}
\title{ntfreq}
\description{
Calculates nucleotide frequency of reads in bam file
}
\section{Methods}{
\describe{

\item{\code{ntfreq(bam_file, ntlength, toRNA = TRUE, count_type = "total")}}{Returns a data frame of nucleotide frequences along length of sequence provided.}

}}
\author{
Diana H.P. Low
}
\keyword{methods}