\name{pingpong-methods}
\docType{methods}
\alias{pingpong-methods}
\alias{pingpong,DataFrame-method}
\title{pingpong}
\description{
piRNA ping-pong analysis of complementary sequences
}
\section{Methods}{
\describe{

\item{\code{pingpong(bam_file)}}{Returns a data.frame object with frequency of overlapping complementary piRNAs.}

}}
\author{
Diana H.P. Low
}
\keyword{methods}