\name{plotDistro-methods}
\docType{methods}
\alias{plotDistro-methods}
\alias{plotDistro,list-method}
\title{plotDistro}
\description{
Plots distribution of reads in the bam file based on length, direction (strand) or location (rname)
}
\section{Methods}{
\describe{

\item{\code{plotDistro(bamlist, type = "qwidth", samplenames = NULL, unique = FALSE, ncounts = 1e+06, norm = FALSE, yname = "Counts per million")}}{Returns a distribution plot.}

}}
\author{
Diana H.P. Low
}
\keyword{methods}