\name{plotFreq-methods}
\docType{methods}
\alias{plotFreq-methods}
\alias{plotFreq,data.frame,logicalORmissing-method}
\title{plotFreq}
\description{
Plots nucleotide frequency generated by \code{\link{ntfreq}}
}
\section{Methods}{
\describe{

\item{\code{plotFreq(freqvector, percentage = TRUE)}}{Returns a frequency bar plot.}

}}
\author{
Diana H.P. Low
}
\keyword{methods}