\name{plotFreq}
\alias{plotFreq}
\title{
plotFreq
}
\description{
Plots nucleotide frequency generated by \code{\link{ntfreq}}
}
\usage{
plotFreq(freqvector, percentage = TRUE)
}

\arguments{
  \item{freqvector}{\code{data.frame} object generated by \code{\link{ntfreq}}
}
  \item{percentage}{Logical value to represent y-axis as percentage or frequency.
}
}
\author{
Diana H.P. Low
}
\seealso{
\code{\link{ntfreq}}
}
\examples{
data(ssviz)
freq<-ntfreq(pctrlbam,ntlength=10)
plotFreq(freq)
}

