\name{plotPP-methods}
\docType{methods}
\alias{plotPP-methods}
\alias{plotPP,list-method}
\title{plotPP}
\description{
Plots the ping-pong frequency of piRNA amplification
}
\section{Methods}{
\describe{

\item{\code{plotPP(pout, samplenames = NULL)}}{Returns the pingpong amplification plot.}

}}
\author{
Diana H.P. Low
}
\keyword{methods}