\name{plotPP}
\alias{plotPP}
\title{plotPP
}
\description{
Plots the ping-pong frequency of piRNA amplification
}
\usage{
plotPP(pout, samplenames = NULL)
}
\arguments{
  \item{pout}{An object of type \code{data.frame} generated by \code{\link{pingpong}}
}
  \item{samplenames}{An object of type character for sample labels.
}
}
\references{
Brennecke J. et al. Cell 128, 1089-1103, March 23, 2007
}
\author{
Diana H.P. Low
}

\seealso{
\code{\link{pingpong}}
}
\examples{
data(ssviz)
pp<-pingpong(pctrlbam)
plotPP(list(pp))
}
