\name{plotRegion-methods}
\docType{methods}
\alias{plotRegion-methods}
\alias{plotRegion,list,character-method}
\title{plotRegion}
\description{
Plots the read density given a chromosome region.
}
\section{Methods}{
\describe{

\item{\code{plotRegion(bamlist, region, howsmooth = 2, ncounts = NULL, samplenames = NULL)}}{Returns the x and y components of the region's reads and plots the density.}

}}
\author{
Diana H.P. Low
}
\keyword{methods}