\name{readBam-methods}
\docType{methods}
\alias{readBam-methods}
\alias{readBam,character-method}
\title{readBam}
\description{
Reads a bam file through RSamtools, and converts it into a data frame of class \code{DataFrame}
}
\section{Methods}{
\describe{

\item{\code{readBam(bam_file, tags = character(0))}}{Returns the bam file contents in a readable dataframe format.}

}}
\author{
Diana H.P. Low
}
\keyword{methods}