\name{ssviz-package}
\alias{ssviz-package}
\alias{ssviz}
\docType{package}
\title{
ssviz
}
\description{
A package for short RNA seq visualization and quantification.
}
\details{
\tabular{ll}{
Package: \tab ssviz\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
Date: \tab 2014-05-08\cr
License: \tab GPL-2\cr
}

}
\author{
Diana H.P. Low
Maintainer: Diana Low <dlow@imcb.a-star.edu.sg>
}
\keyword{ package }